Move RestoreOutputStreamState helper to its own header to be used other places.
